Refactor while let loop to its own module
